+++ src/app/app.config.ts
@@ -6,9 +6,7 @@ import { routes } from './app.routes';
 
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
 import { providePrimeNG } from 'primeng/config';
-import Aura from '@primeuix/themes/aura';
 import { MyPreset } from '../styles';
-
 
 export const appConfig: ApplicationConfig = {
   providers: [
