+++ src/app/features/main/dashboard/trayectoria/trayectoria.html
@@ -1,0 +1,33 @@
+<h3 class="text-3xl font-bold p-4 ">Trayectoria</h3>
+
+<div class="card flex justify-end pb-2">
+    <p-button label="Submit" class="mr-4" />
+    <p-multiselect 
+        [options]="hitos" 
+        [(ngModel)]="selectedHitos" 
+        optionLabel="nombreHito" 
+        placeholder="Seleccione Hitos" 
+        display="chip" 
+        class="w-full md:w-80" 
+    />
+</div>
+
+<div class="space-y-6 w-full py-6 flex flex-col items-center">
+    <!-- Render a chart for each hito -->
+    @for (chart of dataEsteNorte; track chart.nombreHito) {
+        <div class="card bg-white flex justify-center items-center w-[65vw] border rounded-border" id="chartEsteNorte">
+            @if (loading) {
+                <div class="flex justify-center items-center h-[30rem]">
+                    <div class="text-black text-lg">Cargando datos...</div>
+                </div>
+            }
+            @if (!loading) {
+                <p-chart type="scatter" [data]="chart.data" [options]="chart.options" class="h-[30rem] w-[85vw]" />
+            }
+        </div>
+    } @empty {
+        <div class="flex justify-center items-center h-[30rem]">
+            <div class="text-black text-lg">No hay datos disponibles.</div>
+        </div>
+    }
+</div>
