+++ src/app/features/main/dashboard/velocidad/velocidad.html
@@ -1,0 +1,26 @@
+<h3 class="text-3xl font-bold p-4 ">Velocidad</h3>
+
+<div class="card flex justify-end pb-2">
+    <p-multiselect 
+        [options]="hitos" 
+        [(ngModel)]="selectedHitos" 
+        optionLabel="nombreHito"
+        placeholder="Seleccione Hitos" 
+        display="chip" 
+        class="w-full md:w-80" 
+    />
+</div>
+
+<div class="space-y-6 w-[88.5vw] py-6">
+    <!-- Velocidad Media -->
+    <div class="card bg-white border rounded-border" id="chartAverageSpeed">
+        @if (loading) {
+            <div class="flex justify-center items-center h-[30rem]">
+                <div class="text-black text-lg">Cargando datos...</div>
+            </div>
+        }
+        @if (!loading) {
+            <p-chart type="line" [data]="dataAverageSpeed" [options]="optionsAverageSpeed" class="h-[30rem]" />
+        }
+    </div>
+</div>
